requirements.txt
yfinance>=0.2.36
feedparser>=6.0.11
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.3
pytz>=2024.2
lxml>=5.2.0
pandas>=2.2.0
pandas-market-calendars>=4.3.1
